added propbank nodes to unified visualizer
