+++ src/uvi/visualizations/__init__.py
@@ -5,22 +5,20 @@ This module provides classes for creating various visualizations of semantic gra
 including FrameNet and WordNet visualizations, DAG visualizations, taxonomic hierarchies, and interactive plots.
 """
 
-from .Visualizer import Visualizer
 from .InteractiveVisualizer import InteractiveVisualizer
 from .FrameNetVisualizer import FrameNetVisualizer
 from .WordNetVisualizer import WordNetVisualizer
 from .VerbNetVisualizer import VerbNetVisualizer
-from .VerbNetFrameNetWordNetVisualizer import VerbNetFrameNetWordNetVisualizer
+from .UVIVisualizer import UVIVisualizer
 from .PropBankVisualizer import PropBankVisualizer
 from .VisualizerConfig import VisualizerConfig
 
 __all__ = [
-    'Visualizer', 
     'InteractiveVisualizer', 
     'FrameNetVisualizer', 
     'WordNetVisualizer', 
     'VerbNetVisualizer',
-    'VerbNetFrameNetWordNetVisualizer',
+    'UVIVisualizer',
     'PropBankVisualizer',
     'VisualizerConfig'
 ]
